+++ package/src/static-env.ts
