+++ package/src/stubs/dynamic-env/bun.mjs
