+++ package/src/stubs/dynamic-env/dts.d.ts
