+++ package/src/stubs/dynamic-env/node.mjs
