+++ package/src/types.ts
@@ -1,22 +1,6 @@
-import type { AnyZodObject } from "astro/zod";
+import type { AstroGlobal } from "astro";
 
-export type Options = {
-	/**
-	 * Zod schema used to validate your environment variables.
-	 */
-	schema: AnyZodObject;
-	/**
-	 * If set to `true`, generates `.astro/astro-env.d.ts` with types
-	 * based on {schema} and updates `src/env.d.ts`.
-	 *
-	 * @default true
-	 */
-	generateTypes?: boolean;
-	/**
-	 * If set to `true`, generates a `.env.template` with keys based
-	 * on {schema}.
-	 *
-	 * @default false
-	 */
-	generateEnvTemplate?: boolean;
-};
+export type GetEnvVariable = (
+	key: string,
+	options: { locals: AstroGlobal["locals"] },
+) => string | undefined;
