--- package/temp/index.ts
@@ -1,3 +1,0 @@
-import { integration } from "./integration.js"
-
-export default integration;
