--- package/temp/types.ts
@@ -1,6 +1,0 @@
-import type { AstroGlobal } from "astro";
-
-export type GetEnvVariable = (
-	key: string,
-	options: { locals: AstroGlobal["locals"] },
-) => string | undefined;
